+++ modules/dynamic_backends/outputs.tf
@@ -17,11 +17,22 @@
 output "backend_services" {
   description = "The backend service resources."
   value       = google_compute_backend_service.default
+  sensitive   = true // can contain sensitive iap_config
 }
 
 output "external_ip" {
-  description = "The external IP assigned to the global forwarding rule."
+  description = "The external IPv4 assigned to the global fowarding rule."
   value       = local.address
+}
+
+output "external_ipv6_address" {
+  description = "The external IPv6 assigned to the global fowarding rule."
+  value       = local.ipv6_address
+}
+
+output "ipv6_enabled" {
+  description = "Whether IPv6 configuration is enabled on this load-balancer"
+  value       = var.enable_ipv6
 }
 
 output "http_proxy" {